Successfully adds product to database, styling could use a bit of work though
