--- src/components/Css/AddProduct.css
@@ -1,17 +1,0 @@
-label{
-    display: inline-block;
-    clear: left;
-    width: 60%;
-    text-align: right;
-}
-
-
-
-input{
-  width: 350px;
-}
-.Description{
-    height: 250px;
-    width: 350px;
-    vertical-align: top;
-}
