+++ src/components/Products/AddProduct.css
@@ -1,0 +1,8 @@
+.productForm {
+  width:100%;
+  padding: 15px;
+  margin: auto;
+  height: 100%;
+  max-width: 650px;
+}
+
