+++ src/components/Products/FeaturedPage.js
