+++ src/components/Products/SalesPage.js
@@ -1,0 +1,14 @@
+import React from 'react';
+ 
+class SalesPage extends React.Component {
+    render() {
+        return (
+            <div>
+               <h1>SalesPage</h1>
+                <p>Deals</p>
+            </div>
+         );
+    }
+}
+ 
+export default SalesPage;
