+++ src/components/Products/TrendingPage.js
@@ -1,0 +1,14 @@
+import React from 'react';
+ 
+class TrendingPage extends React.Component {
+    render() {
+        return (
+            <div>
+               <h1>TrendingPage</h1>
+                <p>popular items</p>
+            </div>
+         );
+    }
+}
+ 
+export default TrendingPage;
