--- src/components/ThankYouForShoppingPage.js
@@ -1,12 +1,0 @@
-import React from 'react';
-
-class ThankYouForShoppingPage extends React.Component {
-    render(){
-        return(
-            <div className="alert alert-success">
-            <strong>Cool Crepes!</strong> Thank you for shopping!
-            </div>
-        )}
-}
-
-export default ThankYouForShoppingPage;
